Small clarification refactors
