+++ apps/researcher/src/app/[locale]/page.tsx
@@ -8,7 +8,7 @@ import {
   getClientSortBy,
   SearchParams,
 } from '@/lib/search-params';
-import {ClientStore} from 'list-store';
+import {ClientListStore} from 'list-store';
 import {SearchResult} from '@/lib/heritage-fetcher';
 
 // Set the order of the filters
@@ -64,7 +64,7 @@ export default async function Home({searchParams}: Props) {
 
         {searchResult && (
           <>
-            <ClientStore
+            <ClientListStore
               {...{
                 totalCount: searchResult.totalCount,
                 offset: searchResult.offset,
