+++ packages/list-store/package.json
@@ -19,7 +19,7 @@
     "next-intl": "2.13.0-beta.2",
     "react": "18.2.0",
     "zod": "3.21.4",
-    "zustand": "^4.3.7"
+    "zustand": "4.3.7"
   },
   "engines": {
     "node": "18.x"
@@ -28,11 +28,11 @@
     "preset": "ts-jest",
     "testTimeout": 60000,
     "testMatch": [
-      "<rootDir>/src/**/*.test.ts"
+      "<rootDir>/src/**/*.test.ts(x)?"
     ],
     "collectCoverage": true,
     "collectCoverageFrom": [
-      "<rootDir>/src/**/*.ts"
+      "<rootDir>/src/**/*.ts(x)?"
     ]
   }
 }
